refactored tooltip out to component, fixed logic for selecting portfolios and dummy data format
